deployment_2/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
numpy
tensorflow-cpu
scikit-learn==1.2.2